--- .env.development
@@ -1,11 +1,0 @@
-PG_URL=postgresql://postgres:postgres@127.0.0.1:25432/icy_backend_local?sslmode=disable
-DB_HOST=localhost
-DB_PORT=25432
-DB_USER=postgres
-DB_PASS=postgres
-DB_NAME=icy_backend_local
-
-BLOCKCHAIN_BASE_RPC_ENDPOINT=https://base-sepolia-rpc.publicnode.com
-BLOCKCHAIN_BTC_RPC_ENDPOINT=https://blockstream.info/testnet/api
-BLOCKCHAIN_ICY_CONTRACT_ADDR=0x5233E10cc24736F107fEda42ff0157e91Cf1F8b6
-BLOCKCHAIN_BTC_TREASURY_ADDRESS=tb1quvs6zh6c0z77e05vgphks4gp6p7jykvyg6j4e5
