--- internal/oracle/errors.go
@@ -1,7 +1,0 @@
-package oracle
-
-import "errors"
-
-var (
-	ErrBtcSupplyZero = errors.New("btc supply is zero")
-)
